@jwalgran 2022-03-17
I think I found another small bug in the sorting and I believe this allows me to run through the test steps correctly.

By adding `print` statement I found that `active_item_ids` was returning tuples like `(12302,)`. This meant there were never any matches when trying to find the `FacilityListItem` id in the list. Adding `flat=True` appears to correct the issue.

```diff
diff --git a/src/django/api/matching.py b/src/django/api/matching.py
index e388a588..d4637968 100644
--- a/src/django/api/matching.py
+++ b/src/django/api/matching.py
@@ -92,7 +92,7 @@ def get_canonical_items():
 
 def sort_exact_matches(exact_matches, active_item_ids, contributor):
     def sort_order(f):
-        is_active = f['facility_id'] in active_item_ids
+        is_active = f['id'] in active_item_ids
         has_same_contributor = f['source__contributor_id'] == contributor.id
         updated_at = f.get('updated_at', None)
         return (is_active,
@@ -115,7 +115,7 @@ def exact_match_items(messy, contributor):
                             FacilityMatch.MERGED],
                 is_active=True,
                 facility_list_item__source__is_active=True) \
-        .values_list('facility_list_item')
+        .values_list('facility_list_item', flat=True)
 
     results = dict()
 
@@ -131,7 +131,7 @@ def exact_match_items(messy, contributor):
                                              clean_address=clean_address,
                                              country_code=country_code) \
             .exclude(empty_text_fields) \
-            .values('facility_id', 'source__contributor_id', 'updated_at')
+            .values('id', 'facility_id', 'source__contributor_id', 'updated_at')
 
         if len(exact_matches) > 0:
             if len(exact_matches) > 1:
```