@TaiWilkin 2022-03-17
I added the suggested changes and also ran through the `time` and `docker stats` tests. 
For the `time` I'm seeing significantly smaller results than you, interestingly. 
```
real	2m51.825s
user	0m0.593s
sys	0m0.126s
```
For `docker stats`, I see `vagrant_database_1` not going above 45% at the highest, but `vagrant_django_run_5` is staying pretty high (up into the 90s at the highest I saw). I'm not sure whether or not that is a concern. 
(I haven't re-run through the manual sorting tests yet - going to do that next).